finish ability to create an assignment, add a page for displaying information about the assignment
