+++ src/derp/course/templates/course/new_assignment.html
@@ -17,6 +17,8 @@
     <label for="a_description">Description: </label>
     <textarea id="a_description" name="description"></textarea>
     <!-- TODO js plugin for selecting a range, in block-plan-ese -->
+        <label for="a_available">Available: </label><input id="a_available" type="text" name="available" placeholder="2018-12-24"/>
+    <label for="a_due">Due: </label><input id="a_due" type="text" name="due" placeholder="2018-12-25"/>
     <button type="submit" class="btn btn-success">Create!</button>
 </form>
 {% endblock %}
